Fix messed up db setup
