+++ drizzle/0001_seed-data-sources.sql
@@ -1,6 +1,6 @@
--- Custom SQL migration file, put you code below! ---- Custom SQL migration file, put you code below! --
-INSERT INTO data_sources 
-    (id, name)
+-- Custom SQL migration file, put your code below! --
+INSERT INTO data_sources
+(id, name)
 VALUES
     (1, 'Facebook'),
     (2, 'WhatsApp'),
