+++ src/db/schema.ts
@@ -1,4 +1,5 @@
 import * as p from "drizzle-orm/pg-core";
+import {relations} from "drizzle-orm";
 
 export const donationStatus = p.pgEnum("donation_status", ["notstarted", "pending", "complete", "deleted"]);
 
@@ -43,3 +44,46 @@ export const messagesAudio = p.pgTable("messages_audio", {
     lengthSeconds: p.integer("length_seconds"),
     conversationId: p.uuid("conversation_id").notNull().references(() => conversations.id)
 });
+
+export const donationsRelations = relations(donations, ({many}) => ({
+    conversations: many(conversations)
+}));
+
+export const dataSourcesRelations = relations(dataSources, ({ many }) => ({
+    conversations: many(conversations)
+}));
+
+export const conversationsRelations = relations(conversations, ({ one, many }) => ({
+    donation: one(donations, {
+        fields: [conversations.donationId],
+        references: [donations.id]
+    }),
+    dataSource: one(dataSources, {
+        fields: [conversations.dataSourceId],
+        references: [dataSources.id],
+    }),
+    participants: many(conversationParticipants),
+    messages: many(messages),
+    messagesAudio: many(messagesAudio),
+}));
+
+export const conversationParticipantsRelations = relations(conversationParticipants, ({ one }) => ({
+    conversation: one(conversations, {
+        fields: [conversationParticipants.conversationId],
+        references: [conversations.id]
+    })
+}));
+
+export const messagesRelations = relations(messages, ({ one }) => ({
+    conversation: one(conversations, {
+        fields: [messages.conversationId],
+        references: [conversations.id]
+    })
+}));
+
+export const messagesAudioRelations = relations(messagesAudio, ({ one }) => ({
+    conversation: one(conversations, {
+        fields: [messagesAudio.conversationId],
+        references: [conversations.id]
+    })
+}));
